% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_get.R
\name{code_get}
\alias{code_get}
\title{Get company code}
\usage{
code_get(market = "KOSPI")
}
\arguments{
\item{market}{KOSPI is defualt. KOSPI, KOSDAQ, KONEX and all is passible.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get code for korea finance market
}
