% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SHANK}
\alias{SHANK}
\title{Stock prices for the "SHANK" stocks.}
\format{A "tibble" ("tidy" data frame) with 7,580 rows and 8 variables:
\describe{
\item{symbol}{stock ticker symbol}
\item{date}{trade date}
\item{open}{stock price at the open of trading, in won}
\item{high}{stock price at the highest point during trading, in won}
\item{low}{stock price at the lowest point during trading, in won}
\item{close}{stock price at the close of trading, in won}
\item{volume}{number of shares traded}
\item{adjusted}{stock price at the close of trading adjusted for stock splits, in won}
}}
\usage{
SHANK
}
\description{
A dataset containing the daily historical stock prices for the "SHANK" big korea stocks,
"SAMSUNG", "HYUNDAI", "AMOREPACIFIC", "NAVER" and "KAKAO", spanning from 2012-02-08 through 2017-09-07.
}
\keyword{datasets}
