% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_get.R
\name{code_get}
\alias{code_get}
\title{Get company code}
\usage{
code_get(fresh = FALSE)
}
\arguments{
\item{fresh}{get code on internet. Default is FALSE.}
}
\value{
a \link[tibble:tibble-package]{tibble} with market, name, code column.
}
\description{
Get code for korea finance market
}
