% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqk-package.R
\docType{package}
\name{tqk-package}
\alias{tqk}
\alias{tqk-package}
\title{tqk: Get Financial Data in Korea}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Enables the acquisition of Korean financial market data, designed to integrate seamlessly with the 'tidyquant' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/tqk}
  \item \url{https://mrchypark.github.io/tqk/}
  \item Report bugs at \url{https://github.com/mrchypark/tqk/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
