% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqk-package.R
\docType{package}
\name{tqk-package}
\alias{tqk}
\alias{tqk-package}
\title{tqk: Get Financial data in korea with tidyquant package}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Support korea financial data for tidy quant package. 
             especially add stock.prices.korea option in tq_get function. 
             for now, provide tqk_get function instead.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/mrchypark/tqk}
  \item \url{https://mrchypark.github.io/tqk}
  \item Report bugs at \url{http://github.com/mrchypark/tqk/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
