% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tqk_get.R
\name{tqk_get}
\alias{tqk_get}
\title{Get quantitative data from korea in `tibble` format}
\usage{
tqk_get(code, from, to = Sys.Date(), source = c("p", "d", "n"))
}
\arguments{
\item{code}{Stock code.}

\item{from}{Optional for various time series functions. A character string representing a start date in YYYY-MM-DD format.}

\item{to}{Optional for various time series functions. A character string representing a end date in YYYY-MM-DD format.}

\item{source}{not use now.}
}
\description{
Get quantitative data from korea in `tibble` format
}
